+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.3.0
+* Added boolean property `disableTextInput`
+* Switched `touchStart` from React SyntheticEvent to vanilla DOM event handler
+* Added `preventDefault` on touch events to stop unintentional scrolling
+* Added `touchCancel` handler
+
 # 0.2.0
 * Added boolean property `log` to enable logarithmic scale mode
 * Fixed input coercing when `step < 0`